added looks from arif
